Changed the abbreviated state name (ACT) to the full name (Australian Capital Territory) to be consistent with other holiday providers.
